[coordinate] deprecate util/promote_floating_point.hpp
