+++ include/boost/geometry/strategies/geographic/mapping_ssf.hpp
@@ -17,10 +17,10 @@
 
 #include <boost/core/ignore_unused.hpp>
 
+#include <boost/geometry/core/coordinate_promotion.hpp>
 #include <boost/geometry/core/radius.hpp>
 
 #include <boost/geometry/util/math.hpp>
-#include <boost/geometry/util/promote_floating_point.hpp>
 #include <boost/geometry/util/select_calculation_type.hpp>
 
 #include <boost/geometry/strategies/side.hpp>
