+++ include/boost/geometry/strategies/geographic/side.hpp
@@ -16,6 +16,7 @@
 
 #include <boost/geometry/core/cs.hpp>
 #include <boost/geometry/core/access.hpp>
+#include <boost/geometry/core/coordinate_promotion.hpp>
 #include <boost/geometry/core/radian_access.hpp>
 #include <boost/geometry/core/radius.hpp>
 
@@ -32,7 +33,6 @@
 #include <boost/geometry/strategy/geographic/envelope.hpp>
 
 #include <boost/geometry/util/math.hpp>
-#include <boost/geometry/util/promote_floating_point.hpp>
 #include <boost/geometry/util/select_calculation_type.hpp>
 
 namespace boost { namespace geometry
