+++ include/boost/geometry/strategies/spherical/distance_cross_track.hpp
@@ -24,6 +24,7 @@
 
 #include <boost/geometry/core/cs.hpp>
 #include <boost/geometry/core/access.hpp>
+#include <boost/geometry/core/coordinate_promotion.hpp>
 #include <boost/geometry/core/radian_access.hpp>
 #include <boost/geometry/core/tags.hpp>
 
@@ -36,7 +37,6 @@
 #include <boost/geometry/strategies/spherical/intersection.hpp>
 
 #include <boost/geometry/util/math.hpp>
-#include <boost/geometry/util/promote_floating_point.hpp>
 #include <boost/geometry/util/select_calculation_type.hpp>
 
 #ifdef BOOST_GEOMETRY_DEBUG_CROSS_TRACK
